Updated networkmanager

+++ scripts/networkmanager/01-iiserlogin.sh
@@ -2,17 +2,13 @@
 interface="$1"
 status="$2"
 
-possibly_connected_to_iiser() {
-    [ "$interface" = eno1 ] ||
-        { [ "$interface" = wlp5s0 ] && [ "$CONNECTION_ID" = Students ] ;}
-}
-
-case "$status" in
-    down)
-        possibly_connected_to_iiser && systemctl stop iiserlogin.service
-        ;;
-    dhcp4-change|up)
-        [ "$DHCP4_DOMAIN_NAME" = "iiserpune.ac.in" ] &&
-            possibly_connected_to_iiser && systemctl restart iiserlogin.service
-        ;;
-esac
+if [ "$status" = down ] ; then
+    possibly_connected_to_iiser && systemctl stop iiserlogin.service
+elif [ "$interface" = eno1 ] ; then
+    if [ "$status" = up ] || [ "$status" = "dhcp4-change" ] ; then
+        systemctl restart iiserlogin.service
+    fi
+elif [ "$status" = up ] && [ "$interface" = wlp5s0 ] &&
+     [ "$CONNECTION_ID" = Students ] ; then
+    systemctl restart iiserlogin.service
+fi
